Fix win and mac GN failures.
R=scottmg@chromium.org

Review URL: https://codereview.chromium.org/1055103002

Cr-Commit-Position: refs/heads/master@{#323542}

Former-commit-id: 8cd7f113715845b012655f22f7f8cc7fbab960d6

+++ mojo/services/window_manager/window_manager_apptest.cc
@@ -184,9 +184,9 @@ struct FocusedAndActiveViewsCallback {
 
 TEST_F(WindowManagerApplicationTest, GetFocusedAndActiveViewsFailsWithoutFC) {
   EmbedApplicationWithURL(application_impl()->url());
-  uint32 capture_view_id = -1;
-  uint32 focused_view_id = -1;
-  uint32 active_view_id = -1;
+  uint32 capture_view_id = static_cast<uint32>(-1);
+  uint32 focused_view_id = static_cast<uint32>(-1);
+  uint32 active_view_id = static_cast<uint32>(-1);
   base::RunLoop run_loop;
 
   WindowManagerObserverPtr observer;
